webapp/requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
joblib==0.16.0
numpy==1.14.5
sklearn==0.0
pytest==6.0.1
PyYAML==5.3.1
SQLAlchemy==1.3.19
uuid==1.30
